[go/ai] - maintain per-role message counts and per-model usage on chats (#72)
* [malonaz/ai] - add per-role message counts and per-model usage to chat, widen quantity to int64

* [go/ai] - maintain chat message counts and per-model usage

Counts move atomically in SQL from the CreateMessage/DeleteMessage/UndeleteMessage
overrides. The generation rollup now re-reads the chat before writing (retrying
on etag races) instead of reusing the snapshot taken before the model ran.

* [go/ai] - allow model_usages in UpdateChat and cover the chat rollup in the sat

+++ go/ai/ai_service/provider/anthropic/generate_message.go
@@ -212,13 +212,13 @@ func (c *Client) StreamGenerateMessage(
 		case anthropic.MessageStartEvent:
 			modelUsage := &aipb.ModelUsage{Model: request.Model}
 			if variant.Message.Usage.InputTokens > 0 {
-				modelUsage.InputToken = &aipb.ResourceConsumption{Quantity: int32(variant.Message.Usage.InputTokens)}
+				modelUsage.InputToken = &aipb.ResourceConsumption{Quantity: int64(variant.Message.Usage.InputTokens)}
 			}
 			if variant.Message.Usage.CacheReadInputTokens > 0 {
-				modelUsage.InputTokenCacheRead = &aipb.ResourceConsumption{Quantity: int32(variant.Message.Usage.CacheReadInputTokens)}
+				modelUsage.InputTokenCacheRead = &aipb.ResourceConsumption{Quantity: int64(variant.Message.Usage.CacheReadInputTokens)}
 			}
 			if variant.Message.Usage.CacheCreationInputTokens > 0 {
-				modelUsage.InputTokenCacheWrite = &aipb.ResourceConsumption{Quantity: int32(variant.Message.Usage.CacheCreationInputTokens)}
+				modelUsage.InputTokenCacheWrite = &aipb.ResourceConsumption{Quantity: int64(variant.Message.Usage.CacheCreationInputTokens)}
 			}
 			sender.SendModelUsage(ctx, modelUsage)
 
@@ -288,8 +288,8 @@ func (c *Client) StreamGenerateMessage(
 			modelUsage := &aipb.ModelUsage{Model: request.Model}
 			// Anthropic reports thinking_tokens as a subset of output_tokens, but ModelUsage
 			// consumptions must be disjoint since each is priced separately, so split them out.
-			thinkingTokens := int32(variant.Usage.OutputTokensDetails.ThinkingTokens)
-			if outputTokens := int32(variant.Usage.OutputTokens) - thinkingTokens; outputTokens > 0 {
+			thinkingTokens := int64(variant.Usage.OutputTokensDetails.ThinkingTokens)
+			if outputTokens := int64(variant.Usage.OutputTokens) - thinkingTokens; outputTokens > 0 {
 				modelUsage.OutputToken = &aipb.ResourceConsumption{Quantity: outputTokens}
 			}
 			if thinkingTokens > 0 {
